pyproject.toml
[project]
name = "aiogram-dishes"
version = "0.1.0"
description = "This is a telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "google-auth>=2.40.3",
    "ibm-generative-ai>=2.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "google-generativeai",
    "google-cloud-aiplatform",
    "Pillow",
]
